# src/main/kotlin/app/gym/api/response/AddImageResponse.kt
package app.gym.api.response

import java.util.*

data class AddImageResponse(
    val id: UUID
) {
    companion object {
        fun from(id: UUID): AddImageResponse {
            return AddImageResponse(id)
        }
    }
}