# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Product API
  description: Apis to get and manage products
paths:
...
  /api/{image}:
    post:
      tags:
        - ProductImage
      summary: Add image details
      operationId: addImage
      description: Add image details
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddImageResponse"
components:
  schemas:
...
    AddImageResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: id
          example: "550e8400-e29b-41d4-a716-446655440000"